backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
g4f[async]
httpx
pytest
pytest-json-report

